rename test files to match source files and use hyphens
- tests/server/test_image_utils.test.ts -> image-utils.test.ts
- tests/server/test_main.test.ts -> api.test.ts
- tests/eval/classification.eval.test.ts -> wearable-classification.eval.test.ts
- update top-level describe to match source file name (image-utils, api,
  wearable-classification, add, home, schemas, authenticated-image)
